Ifpack2: Fixed link errors for Thyra adapters.
The Thyra adapters' CMakeLists.txt library specification
(TRIBITS_ADD_LIBRARY) for the ifpack2-adapters library did not include
ifpack2 in its DEPLIBS list.  (In fact, it did not have a DEPLIBS list.)
This caused link errors for me with explicit instantiation turned on,
when building with Tpetra, Ifpack2, and Thyra enabled.  I added a
DEPLIBS list with ifpack2 to the TRIBITS_ADD_LIBRARY specification for
ifpack2-adapters, and it fixed the link errors.