Tpetra: VbrMatrix: Simplified VbrDataDist implementation class.
VbrDataDist never needed full DistObject functionality, and never
actually implemented all of DistObject's methods (other than to
throw an exception).  It was only meant to be the source of an
Import or Export to a VbrMatrix, never a target.  Thus, with the
introduction of the new SrcDistObject class, it is natural for
VbrDataDist to inherit from SrcDistObject, rather than DistObject.
This commit implements that change.  It also makes VbrDataDist
inherit from Packable, since it implements that interface already.