Ifpack2: disable overlap if only on one proc
In AdditiveSchwarz, it doesn't make any sense to overlap on one processor.  This was partially
enabled in the Ifpack2::AdditiveSchwarz class, but was missing in the setParameters method.
This change is necessary to be able to use AS as a smoother in a multigrid method, where a coarse
level problem may be created on a single processor.