stk_mesh: modifcation-end optimization: don't shrink-to-fit unmodified connectivity
Shrink to fit causes a full copy of the contents of the vector. If no changes
have been made to a connectivity since the last modification, there is no reason
to pay for this copy.

On this test (run on my blade):
adagio_rtest_performance_p047_cse_dogbone_p047_cse_dogbone.test

Total number of STK-related allocations was reduced from 9367 to 8751.

Time was reduced by about eight percent over five runs on my blade.

For reasons I don't understand, this commit seems to have little impact on
the performance tests when they are run on redsky, providing only a minor
speedup for a handful of tests.

(From Sierra)