NOX/LOCA: Fixed Thyra view of parameter vector
The LOCA abstract interfaces are defined in terms of a simple
ParameterVector class independent from the underlying linear algebra
package in order to achieve generality.

The concrete implementation of the Thyra::Group must convert the data
from the ParameterVector provided by the client and the Stepper to the
subtype of Thyra::VectorBase native to the model evaluator describing
the problem in order to perform the required evaluations.
That is achieved by creating an object implementing Thyra::VectorBase
that is a view of the internal data (a contiguous array) of the
ParameterVector object.

The previous implementation seemed to work flawlessly as long as Epetra
was the linear algebra package, since it relied on the particular
implementation details of the Thyra adapter code for Epetra.
However switching to Tpetra exposed a bug due to faulty assumptions.

The first problem was the use of a non-const view whereas the Thyra
vector is not intended to be used to modify the original ParamVector. In
this particular case, a const view is the correct choice.

The second problem was the wrong assumption that a view object always
remains synchronized with the parent object when the data in the latter
changes. This hypothesis only holds for so-called "direct" views, but
Thyra also allow "detached" views where a temporary copy of the data is
made, leaving the original data free of aliasing. To prevent divergence
between the data in the parent and in the view object, clients are
required to leave the parent unchanged as long as the view exists.

The concrete implementation of Thyra::VectorSpaceBase::createMemberView
used when the underlying linear algebra package is Epetra is the one
from Thyra::DefaultSpmdVectorSpace, which creates "direct" views that
alias the data of the parent object. Therefore, the synchronization
assumption held in that particular case.

However, when Tpetra is the underlying linear algebra engine, there is
(currently) no such specialization for Thyra::TpetraVectorSpace. The
default implementation of Thyra::VectorSpaceBase::createMemberView,
defined in Thyra::VectorSpaceBaseDefault, creates detached views so that
the synchronization assumption is violated, leading to inconsistent
behavior.

The implementation of LOCA::Thyra::Group has been rewritten to avoid
relying on the erroneous hypothesis described above. It is not as
efficient since a new Thyra view must be created every time the parent
LOCA::ParameterVector is updated. However, for an efficient
implementation that allows direct views, the overhead should be small
since there is no bulk data allocation.

A follow-up change to improve the performance for Tpetra would be to
specialize createMemberView in Thyra::TpetraVectorSpace so that it
creates direct views.