Fixing RTOp and Thyra tests for float and complex (Trilinos #5902)
After allowing the re-enabling of float and complex for the SS builds, I was
able to reproduce the build failures shown on the Trilinos CDash server.  They
were the typical std::complex<T> problems of not allowing to compare to an
integral type and not allowing to be multipled by a integral type.  This is
really stupid about std::complex.  I would like to fine the persons that
decided that this was a good idea and have a word with them.  There is nothing
unsafe about upconverting from an integral type to a floating point complex
type (except in the most rare situation).