Amesos: Fix backwards compatibility issue
A recent change to epetra on master removed a using namespace std statement from
the Epetra_config.h header which was causing nearly every file in Epetra to
inadvertantly have the std symbols in the global namespace. This change has
caused a small build error when building against a newer version of Epetra. This
fixes the errors that were reported in the backwards compatibility tests.