PARAVIEW: ParaView Catalyst input deck commands.
The results output section of the input deck now contains
command blocks and line commands to drive the ParaView
Catalyst co-processing library. These command blocks include
blocks for positioning cameras, describing how the mesh is
rendered, and for how the output images are stored.  There are
also blocks to specify visualization operations such as slice
and contour. We are working on detailed PDF documentation of
these commands and how to use them.

The database type has been changed to catalyst from paraview.  The
script name is now catalyst_script. There is also logic for finding
the catalyst dynamic library and the Python module if the Sierra
module is loaded.  The path is constructed using SIERRA_INSTALL_DIR,
SIERRA_SYSTEM, and SIERRA_VERSION.

(From Sierra)