Disable 11.2 backwards compatibility testing
Disabling this testing to reduce the time the nightly testing takes since it is
currently pushing the continuous start to well into the morning. This testing
may be re-enabled lated as a weekend build, but for now the 11.4 backwards
compatibility testing will have to catch the majority of backwards compatibility
issues.