Tpetra: don't store zeros in mat mat product
This commit adds a check for whether the current entry of the left
matrix is zero. If it is, don't bother computing and don't store
the zero value.

This does change the matrix graph, so reuse could be problematic.
However, this reduces storage by about 2x in the first coarse
grid of a MueLu hierarchy.

Build/Test Cases Summary
Enabled Packages: Tpetra
Disabled Packages: MOOCHO
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=618,notpassed=0 (27.58 min)
1) SERIAL_RELEASE => passed: passed=560,notpassed=0 (11.82 min)
2) MPI_SS_DEBUG => passed: passed=475,notpassed=0 (23.40 min)
3) SERIAL_SS_DEBUG => passed: passed=428,notpassed=0 (16.95 min)