Ifpack2: SparseContainer: Simplified gather-scatter for local op.
I also added a new expert mode constructor to Tpetra::MultiVector.
I had thought it would be useful for SparseContainer.  It turned out
that I didn't need it, but it seems like it would be useful for other
reasons, so I left it in place.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=22,notpassed=0 (0.48 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.08 min)