Ifpack2: SparseContainer: Revised public docs, and fixed interface.
This commit removes the unnecessary and optional second argument from
the constructor of SparseContainer, namely the number of vectors
(columns) expected in the input of apply() and weightedApply().  The
latter methods reallocate internal storage on demand anyway, so there is
no need to know in advance how many columns X and Y will have.