Fixed major bug in the handling of symmetry in BKS.
It is possible, because of full orthogonalization and numerical roundoff,
that the Schur factorization of the block upper Hessenberg form
will result in Ritz values with small imaginary part.  If these small complex
Ritz values are ignored in the reordering of the Schur form for restarting
or computing the Ritz vectors, then the Ritz vectors may be wrong
(even though the Ritz values are right).  The BKS solver now detects when
complex Ritz values are computed when the eigenproblem is supposed to be
Hermitian.  This enables the code to sort the Schur form correctly and return
the correct eigenvectors in most cases.  If the complex Ritz values are not the
eigenvalue of interest, then the eigenvectors are computed correctly.  If they
are in the eigenvalues of interest or are on the boundary of the restart subspace,
then the user will be warned that BKS may not return the correct eigenvalues/vectors.