Panzer: Changing how local element ids are computed
These were previously being stored in a mesh field. This was basically
because the mapping from GlobalID->Entity->Field was easy. However,
it becomes more complicated if there are shared elements. Thus the
computation was replaced with a hash table (a boost unordered map).
Because the hash table implementation was "smarter" and checks if the
element has actually been assigned a local ID, a failure was generated
in the SurfaceNodeNormals calculation. This was subsequently overcome
by adding a getElementVertices function that takes the list of mesh
entities.

This commit also adds a getNeighborElements (which returns
the one-ring of elements). This capability has preliminary testing
in the new "tGhosting" unit test.