Ifpack2: Partitioner: Updated public documentation.
The public documentation of Partitioner appears to have been a
copy-and-paste of Ifpack2_Partitioner's documentation.  I updated it for
Ifpack2 and also fixed the included code example.  I haven't tested the
embedded example to see if it compiles, but I figured updating it to a
state of uncertain correctness was better than leaving it in an entirely
wrong state.