Ifpack2: Chebyshev: Improved public documentation.
For some reason, only Ifpack2::Details::Chebyshev (an implementation
detail of Ifpack2::Chebyshev) shows up in the public Doxygen
documentation, not Ifpack2::Chebyshev (which is the interface that users
should actually use).  This commit attempts to address that.  I couldn't
figure out what was wrong, but I tried perturbing the Doxygen a bit to
see if that would help.  I also expanded the documentation of the new
(as of this summer -- thanks to Sarah Osborn of Texas Tech U) clone()
method, and refined its template parameter name to comply with Tpetra
and Ifpack2 conventions.  (Template parameters should be in CamelCase
and should preferably end with "Type", while typedefs that come from
template parameters should be in lowercase with underscore delimiters
and end with "type".)

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=22,notpassed=0 (1.68 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.75 min)