Kokkos: View: Improved public documentation.
In particular, I added an extended explanation of the template
parameters, and an example illustrating how we want users to specify
MemoryTraits ("defer specification until the code that relies on it").

I haven't yet tackled adding Doxygen documentation to the typedefs and
methods of View.  This will be trickier, because the Doxygen class
documentation of View is currently attached to the forward declaration
of View, rather than to the partial specialization(s) where View's
typedefs and methods actually get defined.