Ifpack2: AdditiveSchwarz: Cleaned up code and improved documentation.
I tried to make some sense of this code, which bears marks of a hasty
port from Ifpack.  Please note the "FIXME" annotations I added and
address them if you have a chance.

Separate design discussion:

I have to say that I don't like the second template parameter of
AdditiveSchwarz.  The only reason it is necessary is because of
applyTempl() ("template methods can't be virtual").  This is not a
common use case, even though it relates to one of the original design
goals for Tpetra.  The whole point of abstract interfaces is to favor
run-time polymorphism.  Factoring out the second template parameter and
controlling creation of the local solver entirely through the input
ParameterList would also make it easier to control AdditiveSchwarz
through Stratimikos, since Stratimikos could just pass in a complete
ParameterList (with local solver parameters); it wouldn't need to know
the type of the second template parameter.

One way to make applyTempl() work without the second template parameter
would be to require that users pass in a templated inner solver to
applyTempl().  We could instead add a way to "clone" an
Ifpack2::Preconditioner to support different DomainScalar and
RangeScalar types, analogous to how Tpetra::CrsMatrix::clone() lets
users get a matrix with a different Node type.

Build/Test Cases Summary
Enabled Packages: Ifpack2, Xpetra, Zoltan2
0) MPI_DEBUG => passed: passed=35,notpassed=0 (1.54 min)
1) SERIAL_RELEASE => passed: passed=31,notpassed=0 (0.88 min)
Other local commits for this build/test group: fe9e818