Kokkos: Global and team-local thread-scalable parallel scan for the 'Threads' device.
Use TBB-motivated functor API to allow arbitrary value type and 'join' operator,
and avoid allocating temporary data.

The functor operator is called in two passes:
  Functor::operator()( int iwork , value_type & update , bool final_pass );
On both passes the functor contributes to 'update'.
On the final pass (final_pass == true) the 'update' value is input
as the inclusive scan value and output as the exclusive scan value,
due to the addition of the contributed value.

Threads device team_scan with optional global unordered iter-team global accumulation
for base value of the team scan.