Intrepid: Removed extra ';' in class definition
Removed trailing semicolon from the definition of a macro that
is used to introduce a nested type.

The C++ standard disallows extraneous semicolons inside class
definitions as they correspond to (empty) statements, and some compilers
correctly enforce the rule.

For the sake of uniformity, it is better to let the caller add the
semicolon.