stk_io: change assert to if
*When a nodal field is only declared on specific parts and is used
for restart, zeros are written for the remaining nodes.  When the
field is read in from restart, it should only populate the field
on the parts it is declared.  This change makes it so that the field
data is not written for the nodes that are not declared on parts
instead of failing at an assert.

(From Sierra)