Panzer: Modifying how local ids are computed
This allows a little more flexibility while maintaing decent
performance (and not suffering a frequent v-table lookup). Now
the local ids can be set directly through a method call. This
will allow a deriving DOFManager to set their local ids as
appropriate. This functionality may be more open in the
future, but for now this is sufficient.