Tpetra: Added the first of several "lessons."
I had developed several "lessons" (Tpetra tutorials) on Trilinos' Google
Code wiki.  I've used these lessons to give hands-on Tpetra tutorials,
and have heard that they were useful outside that context as well.  This
commit starts the process of bringing back those lessons to the Trilinos
source repository, so that they are included in the Doxygen
documentation and tested automatically in the check-in and nightly
tests.

This commit adds the first of those lessons, the "Hello world!" example
that shows how to initialize MPI and get a Teuchos::Comm:

http://code.google.com/p/trilinos/wiki/Tpetra_Lesson01_HelloWorld

The text of Lesson 01 lives in example/Lesson01-Init/index.doc.  I
modified Tpetra's Doxyfile (Doxygen configuration file) to include that
file.  It shows up as one of the "Related Pages".  The lesson refers to
three different examples, in three .cpp files in the Lesson01-Init/
directory.  CMake builds and tests these files.  The three .cpp files
each include a link to the text of Lesson 01.

Future commits will bring back the rest of the lessons from the Trilinos
Google Code wiki.