Ifpack2: AdditiveSchwarz: Added comment deprecating 2nd template param.
I plan to deprecate the second template parameter of AdditiveSchwarz.
It adds a whole lot of combinations to Ifpack2's explicit template
instantiation, without adding value (other than making the applyTempl()
method work, an unused method which I removed in a previous commit).
This comment announces the planned deprecation.

I think it might be possible to announce the deprecation syntactially,
through partial template instantiation.  (Deprecate the "most generic"
two-parameter version, but not the partial specialization for a second
argument of Ifpack2::Preconditioner.)  I'll try it out and see if it
works.  I'm worried that some compilers might attach the deprecation
attribute to the class' name instead of to each specialization, which
would cause unexpected compiler warnings that would break
errors-as-warnings builds.  On the other hand, the partial
specialization approach would let me write AdditiveSchwarz so that it
uses Factory to create the subdomain solver, which is how I want to do
it post-deprecation.