Kokkos: CrsMatrix SparseRowView now takes Views, MatVec improvements
Getting a SparseRowView was internally implementated as taking pointers.
Now it takes the views of the matrix instead, this means specialisation
might only need to change the SparseRowView class not the CrsMatrix class.

Matvec with a rank 2 RHS where the second dimension is 1 is now calling
the single RHS vector version, giving a significant speedup on GPUs.

Christian