STK_IO: modifications to get_global functions:
Added optional boolean argument 'abort_if_not_found' which defaults to true.
If true, and the global variable is not found, will throw an exception.
If false and the global variable is not found, will return 'false'
If true/false and the global variable is found, will return 'true'
Requested by Stefan, discussed with conchas2 and nalu developers

Some formatting changes to include file to make it look better
if included in documentation via the latex listings package.

(From Sierra)