KokkosLinAlg:  Tweaks to CrsMatrix for Stokhos embedded scalar types.
Stokhos needs a way to control thread block/team size for MV_multiply()
kernel specializations, so I added a little struct to CrsMatrix to
support this.  This is a simpler case of a general "state" needed by
TPLs, and so this is just a hack until we figure out how to do this
properly.

Also changed the layout of the matrix entries array to LayoutRight,
which is what embedded scalar types will want.  For anything else, it
shouldn't matter.

Fixed the initialization order in one constructor to eliminate warnings.