Ifpack2: Fixed failing tests in my previous commit.
My previous commit shifted creation of single-level preconditioners from
Factory to Details::OneLevelFactory, and also changed AdditiveSchwarz's
default creation of its inner preconditioner to use
Details::OneLevelFactory.  This caused some Ifpack2 tests to fail with a
deference of a null RCP to Tpetra::CrsMatrix.  The issue was that
AdditiveSchwarz initially creates its inner preconditioner with a null
(inner) matrix, but OneLevelFactory was (unnecessarily) asking the
matrix's communicator for its number of processes.  That was
copy-and-paste code from Factory.  OneLevelFactory had no need for the
number of processes, so I removed that code.  That made all of Ifpack2's
tests pass.

This commit will _not_ fix the issue that some Ifpack2 single-level
preconditioners might still object to their constructors being given a
null input matrix.  AdditiveSchwarz does check that its inner
preconditioner implements Details::CanChangeMatrix, but it can only do
that _after_ creation of the inner preconditioner (since the check
involves a dynamic_cast of the inner preconditioner instance).

This issue is straightforward to fix: just make any offending inner
preconditioner implement Details::CanChangeMatrix.  The class
Details::DenseSolver illustrates the general approach.  This also
requires ensuring that the class implements the correct semantics for
initialize() and compute().  Some of Ifpack2's preconditioners (e.g.,
RILUK) are currently broken with respect to those semantics.  In
particular, if the user calls initialize(), it implies that the graph
structure of the matrix may have changed.  This means that the
preconditioner must recompute the "symbolic factorization"; it may not
just check if initialize() has been called before, or if the symbolic
factorization has been computed before.  Furthermore, if a
preconditioner implements CanChangeMatrix, then initialize() and
compute() must throw an exception if the preconditioner's matrix is
null.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=23,notpassed=0 (0.48 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (2.98 min)