Ifpack2: Diagonal: Now implements Details::CanChangeMatrix.
Diagonal now implements Details::CanChangeMatrix.  This makes it fully
legitimate as a subdomain solver for AdditiveSchwarz, or as an inner
preconditioner for any other nested preconditioner.

The only single-level preconditioner left to make implement
Details::CanChangeMatrix is RILUK.  This will be a bit tricky, as the
semantics of the class might not necessarily conform with those of the
other Ifpack2 classes.